Rename: #4 - dto.artist로 artist 관련 dto 이동
