+++ src/main/java/com/teamddd/duckmap/dto/artist/ArtistRes.java
@@ -1,4 +1,6 @@
-package com.teamddd.duckmap.dto;
+package com.teamddd.duckmap.dto.artist;
+
+import com.teamddd.duckmap.dto.ImageRes;
 
 import lombok.Builder;
 import lombok.Getter;
