+++ src/main/java/com/teamddd/duckmap/dto/artist/CreateArtistRes.java
@@ -1,4 +1,4 @@
-package com.teamddd.duckmap.dto;
+package com.teamddd.duckmap.dto.artist;
 
 import lombok.Builder;
 import lombok.Getter;
